Switch to fingerprint userIds.
This makes invites dramatically shorter, and solves the problem of
selecting a stable identifier.

To make this work, Quiver now always connects as an "owner", so any
client that sends a message on a particular server can always receive
a reply through that server.

This change also makes invites freeform objects to allow more compact
encodings.